### 2019-04-14 #1373 @whedon
[PRE REVIEW]: WatchMe: Software for Reproducible Monitoring and Data Collection
**Submitting author:** @vsoch (<a href="http://orcid.org/0000-0002-4387-3819">Vanessa Sochat</a>)
**Repository:** <a href="https://www.github.com/vsoch/watchme" target ="_blank">https://www.github.com/vsoch/watchme</a>
**Version:** 0.0.13
**Reviewers:** @andreysmelter and @crew102
