### 2019-04-15 #1388 @whedon
[REVIEW]: WatchMe: Software for Reproducible Monitoring and Data Collection
**Submitting author:** @vsoch (<a href="http://orcid.org/0000-0002-4387-3819">Vanessa Sochat</a>)
**Repository:** <a href="https://www.github.com/vsoch/watchme" target ="_blank">https://www.github.com/vsoch/watchme</a>
**Version:** 0.0.13

@andreysmelter and @crew102, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
