### 2019-04-15 #1388 @vsoch
Thanks @andreysmelter and @crew102 for offering to review, and @brainstorm for being grandmaster editor! This isn't in the paper (as it's an extra thing) but if/when you are interested, I used this tool with git hooks to show an example of how to (automatically update) some simple plots for a watcher repo -> https://vsoch.github.io/2019/watchme-git-hook/

It will make more sense after you play around a bit :)
