### 2019-04-19 #1388 @vsoch
Why would that be a conflict of interest? I'm definitely okay with it!
