### 2019-04-20 #1388 @arfon
> @brainstorm, is it OK for reviewers to make contributions to the repo they're reviewing or would that amount to some kind of conflict of interest? For example, if I see some spelling mistake in the documentation/paper, can I open a PR to fix it as part of my review?

This is completely fine 👍
