### 2019-04-20 c52c9b1
don't rely on pwd package for getting uname and home dir (closes #19)
