### 2019-04-20 4b23cc7
drop test that assumes we're on linux
