### 2019-04-20 65c1f13
updating get_userhome functions and putting paper lines, one per line
Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>
