+++ paper/paper.md
@@ -17,10 +17,8 @@ bibliography: paper.bib
 
 # Summary
 
-[WatchMe](https://vsoch.github.io/watchme/) is a simple tool to allow for reproducibly watching for changes in one or more 
-web pages, system resources, or any task function that is provided to the library.
-It addresses a problem in research that it's highly challenging to create and share
-reproducible tasks, meaning:
+[WatchMe](https://vsoch.github.io/watchme/) is a simple tool to allow for reproducibly watching for changes in one or more web pages, system resources, or any task function that is provided to the library.
+It addresses a problem in research that it's highly challenging to create and share reproducible tasks, meaning:
 
  1. a configuration file (recipe) stores the parameters for tasks including a function to run, a frequency, and any other necessary variables
  2. the tasks are automatically run at some frequency
@@ -28,93 +26,59 @@ reproducible tasks, meaning:
  4. the results collected can be re-assembled into temporal data structures that are ready for analysis
  5. the entire base (configuration, tasks, and results) can be shared via GitHub, and reproduced by others
 
-With WatchMe, a researcher can easily generate a repository (a watcher) that is configured
-to run one or more tasks at a particular frequency, and automatically commit changes to git.
-If he or she chooses, the repository can be pushed to a version control service like GitHub,
-and the entire configuration and set of tasks is easily reproducible by anyone that uses
-the client to get the repository. Each watcher uses git not only for version control of
-configuration files, but as a temporal database from which the results of the task runs can
-be extracted. Every change to a task within a watcher directory is also recorded via
-git, making the entire setup well documented with minimal to no work needed by the
-researcher.
+With WatchMe, a researcher can easily generate a repository (a watcher) that is configured to run one or more tasks at a particular frequency, and automatically commit changes to git.
+If he or she chooses, the repository can be pushed to a version control service like GitHub, and the entire configuration and set of tasks is easily reproducible by anyone that uses the client to get the repository. Each watcher uses git not only for version control of configuration files, but as a temporal database from which the results of the task runs can be extracted. 
+Every change to a task within a watcher directory is also recorded via git, making the entire setup well documented with minimal to no work needed by the researcher.
 
 ## Background
 
-Reproducible monitoring and data collection for an individual researcher is a challenging task. Typically,
-if a web page or system resource is to be monitored, the researcher must write custom scripts and
-extraction steps, and in the best case scenario, he or she uses version control for the scripts or final
-result. While many online services exist to watch for changes in one or more web pages
-(e.g., see https://visualping.io/ for an example service), these resources are problematic
-for research use. Specifically:
+Reproducible monitoring and data collection for an individual researcher is a challenging task. 
+Typically, if a web page or system resource is to be monitored, the researcher must write custom scripts and extraction steps, and in the best case scenario, he or she uses version control for the scripts or final result. 
+While many online services exist to watch for changes in one or more web pages (e.g., see https://visualping.io/ for an example service), these resources are problematic for research use. Specifically:
 
  1. It's typically the case that you will be charged for more than a few pages
  2. It's not appropriate for a research setting where you would want programmatic parsing
  3. The configuration of your watcher is not reproducible.
 
-Thus, WatchMe is ideal for the individual researcher that does not want to (or cannot)
-pay for a service, and wants to be able to share their monitoring tasks with other
-researchers, such as for a publication or similar. It also allows for collaborative data
-collection, as multiple users can run the equivalent task, have data exported named
-uniquely, and then submit a pull request to combine the data. 
-
[…]
