### 2019-04-23 d99e83b
updating CHANGELOG
Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>
