### 2019-04-23 8290931
updating requirements.txt
Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>
