### 2019-04-23 5237015
updating requirements.txt
Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>
