+++ requirements.txt
@@ -1,4 +1,3 @@
 python-crontab>=2.3.6
 configparser>=3.5.3
 requests>=2.21.0
-
