### 2019-04-26 #1388 @andreysmelter
Hi @brainstorm @vsoch
I just have a quick question if I have a minor revisions/comments where do I write them? Trying to finalize review over this weekend.

Thank you.
