### 2019-04-26 #1388 @vsoch
hey @andreysmelter this is great news!

For specific questions or edits that need to be made, what I've typically done is open up an issue for each. It might seem like overkill, but it's easier for both author and reviewer to keep track of what has been fixed, and what not. What I would likely do is reference each issue in one or more PRs to address them.
