### 2019-04-27 a8fb379
Chore: make logo center
