### 2019-04-30 #1388 @andreysmelter
Hi @brainstorm @vsoch
I have completed my review and raised several issues that I have found against the watchme repo. They have already been fixed and merged. Everything else looks excellent to me and ready to go to the next step (acceptance/publishing?).

Thank you.
