### 2019-05-02 cb250d1
Added different config templates for exporters.
