+++ watchme/config/templates/pushgateway/watchme.cfg
@@ -1,0 +1,6 @@
+[watcher]
+active = false
+type = urls
+
+[pushgateway]
+url = localhost:9091
