### 2019-05-02 cd3eeac
Added the exporter option to the create command.
