+++ watchme/config/__init__.py
@@ -26,16 +26,16 @@ import os
 
 # CONFIG TEMPLATES #############################################################
 
-def get_configfile_template(watcher_type=None):
+def get_configfile_template(exporter=None):
     '''return the full path to the default configuration file
     '''
-    return _get_config('watchme.cfg', watcher_type)
+    return _get_config('watchme.cfg', exporter)
 
-def _get_config(name, watcher_type):
+def _get_config(name, exporter):
     '''shared function to return a file in the config directory
     '''
-    watcher_path = watcher_type or ''
-    return os.path.abspath(os.path.join(get_installdir(), 'config', watcher_path, name))
+    exporter_path = exporter or ''
+    return os.path.abspath(os.path.join(get_installdir(), 'config', 'templates', exporter_path, name))
 
 
 # ACTIVE CONFIG FILES ##########################################################
@@ -72,7 +72,7 @@ def read_config(filename):
 
 # WATCHER CONFIG ###############################################################
 
-def generate_watcher_config(path, watcher_type=None):
+def generate_watcher_config(path, watcher_type=None, exporter=None):
     '''generate a watcher config, meaning a watcher folder in the watchme
        base folder.
 
@@ -81,7 +81,7 @@ def generate_watcher_config(path, watcher_type=None):
        path: the path to the watcher repository
     '''
     check_exists(path)
-    configfile = get_configfile_template()
+    configfile = get_configfile_template(exporter)
     watcher_config = os.path.join(path, 'watchme.cfg')
     if not os.path.exists(watcher_config):
         bot.info('Generating watcher config %s' % watcher_config)
