### 2019-05-04 f7ea250
Added a regex option when performing a get_task
