### 2019-05-05 #1388 @vsoch
- [Released via Zenodo](https://zenodo.org/record/2667656#.XM7-j3VKjoo)
 - [Released on PyPi](https://pypi.org/project/watchme/)
