+++ watchme/watchers/__init__.py
@@ -8,7 +8,7 @@ with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 '''
 
-from watchme.logger import ( bot, RobotNamer )
+from watchme.logger import (bot, RobotNamer)
 from watchme.version import __version__
 from watchme.defaults import (
     WATCHME_BASE_DIR,
@@ -70,12 +70,12 @@ import sys
 
 class Watcher(object):
 
-    repo=None
-    configfile=None
+    repo = None
+    configfile = None
 
-    def __init__(self, name=None, 
-                       base=None, 
-                       create=False, **kwargs):
+    def __init__(self, name=None,
+                 base=None,
+                 create=False, **kwargs):
         '''the watcher base loads configuration files for the user (in $HOME)
            and module, and then stores any arguments given from the caller
 
@@ -93,7 +93,6 @@ class Watcher(object):
 
         # Load the configuration
         self.load_config()
-
 
     def _set_base(self, base=None, create=False):
         ''' set the base for the watcher, ensuring that it exists.
@@ -114,9 +113,10 @@ class Watcher(object):
         # If the watcher doesn't exist and we need to create:
         if not os.path.exists(self.repo) or not os.path.exists(self.configfile):
             if create is True:
-                create_watcher(self.name)   
+                create_watcher(self.name)
             else:
-                bot.exit('Watcher %s does not exist. Use watchme create.' % self.name)
+                bot.exit(
+                    'Watcher %s does not exist. Use watchme create.' % self.name)
 
 
 # Config
@@ -124,7 +124,6 @@ class Watcher(object):
     def save(self):
         '''save the configuration to file.'''
         write_config(self.configfile, self.config)
-
 
     def edit_task(self, name, action, key, value=None):
         '''edit a task, meaning doing an addition (add), update (update), or
@@ -139,7 +138,7 @@ class Watcher(object):
         '''
 
         if not self.has_task(name):
-            bot.exit('%s is not a task defined by %s' %(name, self.name))
+            bot.exit('%s is not a task defined by %s' % (name, self.name))
 
         if action not in ['update', 'add', 'remove']:
             bot.exit('Action must be update, add, or remove')
@@ -149,7 +148,7 @@ class Watcher(object):
 
         # Add, and it doesn't exist so it's okay
         if action == "add" and key not in self.config[name]:
-            bot.info('Adding %s:%s to %s' %(key, value, name))
+            bot.info('Adding %s:%s to %s' % (key, value, name))
             self.set_setting(name, key, value)
 
         # Already exists, encourage user to update
@@ -158,7 +157,7 @@ class Watcher(object):
 
         # Update, and it's a valid choice
         elif action == 'update' and key in self.config[name]:
-            bot.info('Updating %s to %s in %s' %(key, value, name))
+            bot.info('Updating %s to %s in %s' % (key, value, name))
             self.set_setting(name, key, value)
 
         # Update, and it's not a valid choice
@@ -167,14 +166,13 @@ class Watcher(object):
 
         # Remove, and it's a valid choice
         elif action == "remove" and key in self.config[name]:
-            bot.info('Removing %s' % key )
+            bot.info('Removing %s' % key)
             del self.config[name][key]
 
         # Remove, and it's not a valid choice
         elif action == "remove" and key not in self.config[name]:
             bot.exit('%s is not found in config, cannot be removed.' % key)
         self.save()
-
 
     def has_section(self, name):
         '''returns True or False to indicate if the watcher has a specified
@@ -187,9 +185,8 @@ class Watcher(object):
         self.load_config()
         if name in self.config._sections:
             return True
-        bot.warning('%s not found for watcher %s' %(name, self.name))
[…]
