+++ watchme/config/templates/pushgateway/watchme.cfg
@@ -2,5 +2,6 @@
 active = false
 type = urls
 
-[pushgateway]
+[exporter-pushgateway]
 url = localhost:9091
+active = true
