### 2019-05-05 a9755e2
Infering exporter instance given its declared type
