+++ watchme/config/templates/pushgateway/watchme.cfg
@@ -4,4 +4,5 @@ type = urls
 
 [exporter-pushgateway]
 url = localhost:9091
+type = pushgateway
 active = true
