### 2019-05-05 a2d724f
Helper function to retrieve the exporter type
