### 2019-05-05 5d9168e
Removed exporter specific function from task classes
