### 2019-05-05 cce9297
Catching exceptions thrown when exporters are not available.
