### 2019-05-05 1d4b890
Handling regex on get_url_selection + error message when exporter does not exist
