+++ watchme/config/__init__.py
@@ -35,7 +35,14 @@ def _get_config(name, exporter):
     '''shared function to return a file in the config directory
     '''
     exporter_path = exporter or ''
-    return os.path.abspath(os.path.join(get_installdir(), 'config', 'templates', exporter_path, name))
+
+    template_path = os.path.join(get_installdir(), 'config', 'templates', exporter_path, name)
+
+    if os.path.exists(template_path):
+        return os.path.abspath(template_path)
+    else:
+        bot.info('The exporter specified does not exist : %s. The task was created with no exporters. ' % exporter)
+        return os.path.abspath(os.path.join(get_installdir(), 'config', 'templates', name))
 
 
 # ACTIVE CONFIG FILES ##########################################################
