### 2019-05-06 dd31e19
adding AUTHORS file
Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>
