+++ .github/AUTHORS.md
@@ -1,0 +1,7 @@
+# Maintainers
+
+ - [@vsoch](https://www.github.com/vsoch)
+
+# Contributors
+
+ - [@SCHKN](https://www.github.com/SCHKN)
