### 2019-05-06 3a08ca4
initial bulk changes working on exporter
Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>
