+++ watchme/watchers/data.py
@@ -21,8 +21,7 @@ from watchme.command import (
 import os
 import json
 
-# Data Exports
-
+# Default (git) Data Exports
 
 def export_dict(self, task,
                       filename, 
@@ -86,3 +85,213 @@ def export_dict(self, task,
         result['dates'].append(git_date(repo=repo, commit=commit))
         result['commits'].append(commit)
     return result
+
+
+# Exporter Functions
+
+    # Add the exporter, optionally with tasks                  
+    watcher.add_exporter(exporter=exporter,
+                         exporter_type=exporter_type,
+                         params=params,
+                         tasks=tasks,
+                         force=args.force,
+                         active=args.active)
+
+
+def add_exporter(self, name, exporter_type, params, tasks, force=False, active="true"):
+    '''add an exporter, meaning an extra plugin to push data to a remote.
+
+       Parameters
+       ==========
+       name: the name of the exporter to add, should start with exporter-
+       exporter_type: must be in WATCHME_EXPORTERS, meaning a client exists
+       params: list of parameters to be validated (key@value)
+       tasks: list of task names to add the exporter to
+       force: if task already exists, overwrite
+       active: add the task as active (default "true")
+    '''
+    # Check again, in case user calling from client
+    if not name.startswith('exporter'):
+        bot.exit('Exporter name must start with "exporter" (e.g., exporter-grafana)')
+
+    # Ensure it's a valid type
+    if exporter_type not in WATCHME_EXPORTERS:
+        bot.exit('%s is not a valid type: %s' % WATCHME_EXPORTERS)
+
+    # Validate variables provided for task
+    if task_type.startswith('pushgateway'):
+        from watchme.exporters.pushgateway import Exporter
+
+    else:
+        bot.exit('exporter_type %s not installed' % exporter_type)
+
+    # Convert list to dictionary
+    params = self._get_params_dict(params)
+
+    # Creating the exporter will validate parameters
+    exporter = Exporter(name, params=params)
+
+    # Exit if the exporter is not valid
+    if not exporter.valid:
+        bot.exit('%s is not valid, will not be added.' % exporter)
+
+    # Write to file (all tasks get active = True added, and type)
+    self._add_exporter(exporter, force, active, tasks)
+
+
+
+def _add_exporter(self, exporter, force=False, active='true', tasks=[]):
+    '''add a new exporter to the watcher, meaning we:
+
+       1. Check first that the task doesn't already exist (if the task
+          exists, we only add if force is set to true)
+       2. Validate the task (depends on the task)
+       3. write the task to the helper config file, if valid.
+
+       Parameters
+       ==========
+       task: the Task object to add, should have a name and params and
+             be child of watchme.tasks.TaskBase
+       force: if task already exists, overwrite
+       active: add the task as active (default "true")
+    '''
+    self.load_config()
+
+    if active not in ["true", "false"]:
+        bot.exit('Active must be "true" or "false"')
+
+    # Don't overwrite a section that already exists
+    if exporter.name in self.config.sections():
+        if not force:
+            bot.exit('%s exists, use --force to overwrite.' % task.name)
+        self.remove_section(exporter.name, save=False)
+
+    # Add the new section
+    self.config[exporter.name] = exporter.export_params(active=active)
+    self.print_section(exporter.name)
+
+    # Add the exporter to any tasks
+    for task in tasks:
+        if task in self.config.sections():
+            exporters = self.get_setting(task, "exporters", default=[])
+            if exporter.name not in exporters:
+                exporters.append(exporter.name)
+                self.set_setting(task, "exporters", exporters)
+        else:
[…]
