+++ watchme/watchers/settings.py
@@ -97,8 +97,8 @@ def print_section(self, section):
         bot.exit('%s is not a valid section.' % section)
 
 def get_setting(self, section, name, default=None):
-    '''return a setting from the environment (first priority) and then
-       secrets (second priority) if one can be found. If not, return None.
+    '''return a setting from the config, if defined. Otherwise return
+       default (None or set by user)
 
        Parameters
        ==========
@@ -118,6 +118,34 @@ def get_setting(self, section, name, default=None):
 
     return setting
 
+
+def has_setting(self, section, name):
+    '''return a boolean if a config has a setting (or not)
+
+       Parameters
+       ==========
+       section: the section in the config, defaults to self.name
+       name: they key (index) of the setting to look up
+    ''' 
+    self.load_config()
+
+    exists = False
+    if section in self.config:
+         if name.lower() in self.config[section]:
+            exists = True
+
+    return exists
+
+
+def has_section(self, section):
+    '''return a boolean if a config has a section (e.g., a task or exporter)
+
+       Parameters
+       ==========
+       section: the section in the config
+    ''' 
+    self.load_config()
+    return section in self.config
 
 
 def set_setting(self, section, key, value):
@@ -145,6 +173,6 @@ def get_section(self, name):
     '''
     section = None
     self.load_config()
-    if name in self.config.sections():
+    if name in self.config:
         section = self.config[name]
     return section
