### 2019-05-06 67f5221
tweaks in setup.py (still not tested)
Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>
