+++ watchme/command/__init__.py
@@ -26,5 +26,7 @@ from .commit import (
 )
 from .utils import (
     get_watchers,
-    list_watcher
+    list_exporters,
+    list_watcher,
+    list_watcher_types
 )
