+++ watchme/watchers/urls/helpers.py
@@ -9,6 +9,7 @@ with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
 '''
 
 import os
+import re
 import tempfile
 import requests
 
