+++ watchme/watchers/data.py
@@ -9,6 +9,7 @@ with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
 '''
 
 from watchme.logger import bot
+from watchme.defaults import WATCHME_EXPORTERS
 from watchme.utils import ( 
     which, 
     get_user
@@ -16,10 +17,12 @@ from watchme.utils import (
 from watchme.command import (
     get_commits,
     git_show,
-    git_date
+    git_date,
+    git_commit
 )
-import os
 import json
+import os
+import re
 
 # Default (git) Data Exports
 
@@ -87,16 +90,15 @@ def export_dict(self, task,
     return result
 
 
+    # Push to the exporter (not the default git)
+    result = watcher.push(task=task,
+                          exporter=exporter,
+                          name=name,
+                          export_json=args.json,
+                          base=args.base)
+
+
 # Exporter Functions
-
-    # Add the exporter, optionally with tasks                  
-    watcher.add_exporter(exporter=exporter,
-                         exporter_type=exporter_type,
-                         params=params,
-                         tasks=tasks,
-                         force=args.force,
-                         active=args.active)
-
 
 def add_exporter(self, name, exporter_type, params, tasks, force=False, active="true"):
     '''add an exporter, meaning an extra plugin to push data to a remote.
@@ -114,30 +116,51 @@ def add_exporter(self, name, exporter_type, params, tasks, force=False, active="
     if not name.startswith('exporter'):
         bot.exit('Exporter name must start with "exporter" (e.g., exporter-grafana)')
 
-    # Ensure it's a valid type
-    if exporter_type not in WATCHME_EXPORTERS:
-        bot.exit('%s is not a valid type: %s' % WATCHME_EXPORTERS)
+    self.load_config()
 
-    # Validate variables provided for task
-    if task_type.startswith('pushgateway'):
-        from watchme.exporters.pushgateway import Exporter
+    # If it already exists and the user isn't recreating, don't add it again.
+    if name in self.config.sections() and exporter_type == None:
+        exporter = self.get_exporter(name)
 
+    # Otherwise, this is creation of a new exporter
     else:
-        bot.exit('exporter_type %s not installed' % exporter_type)
 
-    # Convert list to dictionary
-    params = self._get_params_dict(params)
+        # Ensure it's a valid type
+        if exporter_type not in WATCHME_EXPORTERS:
+            bot.exit('%s is not a valid type: %s' % WATCHME_EXPORTERS)
 
-    # Creating the exporter will validate parameters
-    exporter = Exporter(name, params=params)
+        # Validate variables provided for task
+        if exporter_type.startswith('pushgateway'):
+            from watchme.exporters.pushgateway import Exporter
+
+        else:
+            bot.exit('exporter_type %s not installed' % exporter_type)
+
+        # Convert list to dictionary
+        params = self._get_params_dict(params)
+
+        # Creating the exporter will validate parameters
+        exporter = Exporter(name, params=params)
+
 
     # Exit if the exporter is not valid
     if not exporter.valid:
-        bot.exit('%s is not valid, will not be added.' % exporter)
+        bot.exit('%s is not valid, will not be added.' % exporter.name)
 
-    # Write to file (all tasks get active = True added, and type)
-    self._add_exporter(exporter, force, active, tasks)
+    # Add the exporter to the config
+    if exporter_type != None:
+        self._add_exporter(exporter, force, active)
 
+    # Add tasks to it
+    for task in tasks:
+        self.add_task_exporter(task, exporter.name)
+
+    # Save all changes
+    self.save()
+
+    # Commit changes
+    git_commit(repo=self.repo, task=self.name,
[…]
