### 2019-05-07 c180c0c
add dummy push function, need to think about it
Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>
