### 2019-05-07 a250137
updating watcher task functions to be cleaner, inspired by other PR
Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>
