+++ watchme/watchers/__init__.py
@@ -8,7 +8,7 @@ with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 '''
 
-from watchme.logger import ( bot, RobotNamer )
+from watchme.logger import (bot, RobotNamer)
 from watchme.version import __version__
 from watchme.defaults import (
     WATCHME_BASE_DIR,
@@ -32,12 +32,14 @@ from .data import (
 
 from .settings import (
     get_setting,
-    set_setting,
     get_section,
+    has_setting,
+    has_section,
     print_section,
     print_add_task,
     remove_setting,
-    remove_section
+    remove_section,
+    set_setting
 )
 
 from .schedule import (
@@ -114,7 +116,7 @@ class Watcher(object):
         # If the watcher doesn't exist and we need to create:
         if not os.path.exists(self.repo) or not os.path.exists(self.configfile):
             if create is True:
-                create_watcher(self.name)   
+                create_watcher(self.name)
             else:
                 bot.exit('Watcher %s does not exist. Use watchme create.' % self.name)
 
@@ -139,7 +141,7 @@ class Watcher(object):
         '''
 
         if not self.has_task(name):
-            bot.exit('%s is not a task defined by %s' %(name, self.name))
+            bot.exit('%s is not a task defined by %s' % (name, self.name))
 
         if action not in ['update', 'add', 'remove']:
             bot.exit('Action must be update, add, or remove')
@@ -149,7 +151,7 @@ class Watcher(object):
 
         # Add, and it doesn't exist so it's okay
         if action == "add" and key not in self.config[name]:
-            bot.info('Adding %s:%s to %s' %(key, value, name))
+            bot.info('Adding %s:%s to %s' % (key, value, name))
             self.set_setting(name, key, value)
 
         # Already exists, encourage user to update
@@ -158,7 +160,7 @@ class Watcher(object):
 
         # Update, and it's a valid choice
         elif action == 'update' and key in self.config[name]:
-            bot.info('Updating %s to %s in %s' %(key, value, name))
+            bot.info('Updating %s to %s in %s' % (key, value, name))
             self.set_setting(name, key, value)
 
         # Update, and it's not a valid choice
@@ -167,7 +169,7 @@ class Watcher(object):
 
         # Remove, and it's a valid choice
         elif action == "remove" and key in self.config[name]:
-            bot.info('Removing %s' % key )
+            bot.info('Removing %s' % key)
             del self.config[name][key]
 
         # Remove, and it's not a valid choice
@@ -198,7 +200,7 @@ class Watcher(object):
            it will be written with a default active status set to false.
         '''
         if not hasattr(self, 'config'):
-           
+
             # Load the configuration file if it exists (will exit if not found)
             if self.configfile != None:
                 self.config = read_config(self.configfile)
@@ -224,7 +226,7 @@ class Watcher(object):
         for pair in pairs:
             if "@" not in pair:
                 bot.exit('incorrectly formatted param, must be key@value')
-            key,value = pair.split('@', 1)
+            key, value = pair.split('@', 1)
             key = key.lower()
 
             # All tasks are not allowed to have default params
@@ -249,8 +251,8 @@ class Watcher(object):
            params: list of parameters to be validated (key@value)
            force: if task already exists, overwrite
            active: add the task as active (default "true")
-           
         '''
+
         # Check again, in case user calling from client
         if not task.startswith('task'):
             bot.exit('Task name must start with "task" (e.g., task-reddit)')
@@ -322,7 +324,8 @@ class Watcher(object):
             git_add(self.repo, task.name)
 
         # Commit changes
[…]
