+++ watchme/command/__init__.py
@@ -26,5 +26,6 @@ from .commit import (
 )
 from .utils import (
     get_watchers,
-    list_watcher
+    list_watcher,
+    list_watcher_types
 )
