### 2019-05-08 fc88256
adding results watcher - first from the environment
Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>
