### 2019-05-08 b7a1cf9
updating watchme docs
Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>
