### 2019-05-09 #1388 @vsoch
All set!

https://zenodo.org/record/2667656#.XNQjEnVKjoo

I also added the orcid id, and license.
